Break out and refactor profile upgrading tests; r=canaltinova
The upgrading tests were hard to find, and inconsistent with the current
style of jest testing. This PR collects the upgrades into their own
file, and adds the more typical describe/it blocks that make the tests
run and fail in a more granular fashion.
